+++ src/py42/exceptions.py
@@ -176,8 +176,8 @@ class Py42CloudAliasLimitExceededError(Py42BadRequestError):
         super(Py42BadRequestError, self).__init__(exception, message)
 
 
-class Py42UserDoesNotExist(Py42NotFoundError):
-    """An exception raised when user does not exist in the detection list."""
+class Py42UserNotOnListError(Py42NotFoundError):
+    """An exception raised when user is not in the detection list."""
 
     def __init__(self, exception, message=None):
         super(Py42NotFoundError, self).__init__(exception, message)