+++ src/py42/services/detectionlists/high_risk_employee.py
@@ -1,6 +1,6 @@
 from py42.exceptions import Py42BadRequestError
 from py42.exceptions import Py42NotFoundError
-from py42.exceptions import Py42UserDoesNotExist
+from py42.exceptions import Py42UserNotOnListError
 from py42.services import BaseService
 from py42.services.detectionlists import _DetectionListFilters
 from py42.services.detectionlists import _PAGE_SIZE
@@ -95,10 +95,10 @@ class HighRiskEmployeeService(BaseService):
             return self._connection.post(uri, json=data)
         except Py42NotFoundError as err:
             message = (
-                "User {} is not currently on the high-risk-employee "
+                "User with ID '{}' is not currently on the high-risk-employee "
                 "detection list.".format(user_id)
             )
-            raise Py42UserDoesNotExist(err, message=message)
+            raise Py42UserNotOnListError(err, message=message)
 
     def get(self, user_id):
         """Gets user information.