+++ tests/services/detectionlists/test_high_risk_employee.py
@@ -234,16 +234,16 @@ class TestHighRiskEmployeeClient(object):
     def test_remove_raises_error_when_user_id_does_not_exist(
         self,
         user_context,
-        mock_connection_post_failure,
+        mock_post_not_found_session,
         mock_detection_list_user_client,
     ):
         high_risk_employee_client = HighRiskEmployeeService(
-            mock_connection_post_failure, user_context, mock_detection_list_user_client
+            mock_post_not_found_session, user_context, mock_detection_list_user_client
         )
         user_id = "942897397520289999"
         with pytest.raises(Py42NotFoundError) as err:
             high_risk_employee_client.remove(user_id)
-        assert "User {} is not currently on the high-risk-employee detection list.".format(
+        assert "User with ID '{}' is not currently on the high-risk-employee detection list.".format(
             user_id
         ) in str(
             err.value