+++ CHANGELOG.md
@@ -18,8 +18,7 @@ how a consumer would use the library (e.g. adding unit tests, updating documenta
 
 - Added new exception `Py42CloudAliasLimitExceededError` to throw if `add_cloud_alias()` throws `400` and body contains
 reason `Cloud usernames must be less than or equal to`.
-- Added new exception `Py42UserNotAssignedToRule` to throw error if `alert_rules.remove_user` throws `400` when `user_id` is not linked to the `rule_id`.
-- Added new exception `Py42UserDoesNotExist` to throw error if `remove()` throws `400` when user does not exist in the detection list.
+- Added new exception `Py42UserNotOnListError` to throw error if `remove()` throws `404` when user does not exist in the detection list.
 
 ## 1.9.0 - 2020-10-02